# docs/api/project-settings.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project settings — review S03/S10",
    "version": "0.10.1"
  },
  "paths": {
...
    "/api/v1/projects/{projectId}": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "getProject",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateProject",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/handbook": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listHandbookResources",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createHandbookResource",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResource"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/handbook/{resourceId}": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "resourceId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "getHandbookResource",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDetail"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/handbook/{resourceId}/revisions": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "resourceId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "post": {
        "operationId": "appendHandbookRevision",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceRevision"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TMS_SESSION"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "archived": {
            "type": "boolean"
          },
          "projectRole": {
            "type": "string",
            "maxLength": 16
          }
        },
        "required": []
      },
...
      "UpdateProject": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
...
      "ResourceRevision": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "revisionNo": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "contentHtml": {
            "type": "string",
            "maxLength": 10000,
            "description": "Văn bản thuần; không thực thi HTML."
          },
          "visibility": {
            "type": "string",
            "enum": [
              "INTERNAL"
            ]
          },
          "editedBy": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": []
      },
      "CreateResource": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "resourceType": {
            "type": "string",
            "enum": [
              "GUIDE",
              "SPEC_REFERENCE",
              "TEST_ACCOUNT_REFERENCE"
            ]
          },
          "contentHtml": {
            "type": "string",
            "maxLength": 10000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "INTERNAL"
            ]
          }
        },
        "required": [
          "code",
          "name",
          "resourceType",
          "contentHtml"
        ]
      },
      "CreateResourceRevision": {
        "type": "object",
        "properties": {
          "contentHtml": {
            "type": "string",
            "maxLength": 10000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "INTERNAL"
            ]
          },
          "expectedCurrentRevisionId": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        "required": [
          "contentHtml",
          "expectedCurrentRevisionId"
        ]
      },
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "resourceType": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "archived": {
            "type": "boolean"
          },
          "currentRevision": {
            "$ref": "#/components/schemas/ResourceRevision"
          }
        },
        "required": []
      },
      "ResourceDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Resource"
          },
          {
            "type": "object",
            "properties": {
              "history": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceRevision"
                }
              }
            },
            "required": []
          }
        ]
      },
...
      "ProjectDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "archived": {
            "type": "boolean"
          },
          "membersCount": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      }
    }
  }
}